requirements.txt
tensorflow==2.19.0
numpy
pandas
scikit-learn
matplotlib
joblib
flask
flask-cors
Pillow
